Init
This is the GitHub init, though this is already during upgrades for version 3 of the project. I just now chose to upload, so thanks for nothing past me...
I've reworked the UI, made new animations, a new tracker in 3D, and currently just testing it all thoroughly before I'll attempt to deploy it.