Animation & Help
I've made a new start-up animation, but it doesn't work on the HoloLens. I'll have to look into that. Meanwhile, I'll make a video at home showcasing it in Unity.
I've also added some descriptive text to the tutorial parts, which can be toggled on and off afterwards for help.
And I've recorded some new videos on the HoloLens showcasing some of the newer features!